+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.0
+
+- Test Tabbar features With PageView (horizontal)
+- Test Tabbar features With PageView (vertical)
+- Fix some bug of Tabbar Indicator Position
+- Support Tarbar scroll pinned mode
+
 ## 1.0.0
 
 - Initial repository